Remove the obsolete file upfirdn.m
This file is now superseeded by src/upfirdn.cc.  In normal situations,
keeping both upfird.m and upfirdn.oct in the package will not hurt.
However, there may be cases where the former takes preference on the later
in the path.  In this case, tests of resample and fracshift will fail.


git-svn-id: svn://svn.code.sf.net/p/octave/code/trunk/octave-forge@5862 fc592f45-0881-4044-91ff-c96495a54512