ga: restructure unit tests, fixing bug #3287917
The related Debian bug is #622929.

Convert some unit tests to proper unit tests. Move other unit
tests to ga_demo (they will be converted to demos in the future),
as that code is more suitable for demos than for unit tests (the
result of the ga function is not deterministic by design).

http://sourceforge.net/tracker/?func=detail&aid=3287917&group_id=2888&atid=102888
http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=622929

git-svn-id: svn://svn.code.sf.net/p/octave/code/trunk/octave-forge@9896 fc592f45-0881-4044-91ff-c96495a54512