units: support for units without --compact and --one-line options (bug #38270)
* units.m: on r12290, the call to units was changed to use the compact and
one-line option to ease parsing. This commit recovers part of the old
code to support versions of GNU units pre 1.86 (released on 10 Nov 2006 before
addition of these options) and others that lack this options. In addition,
checks if the units program available is indeed GNU units or some other
implementation.



git-svn-id: svn://svn.code.sf.net/p/octave/code/trunk/octave-forge@12298 fc592f45-0881-4044-91ff-c96495a54512