Replace deprecated usage of symbol_table::varref() with symbol_table::assign().
* unpackfields.cc: symbol_table::varref() has been deprecated in Octave 3.8.
Replaced with a call to symbol_table::asign(). Added tests and fix
documentation to make it clear that input is argument must be the struct
name, not the struct variable.
* packfields.cc: add tests and fix documentation to make it clear that input
is variable and struct names, not the variables and struct themselves.
Remove workaround bug for Octave 3.2.4 since general package is dependent on
Octave 3.4.3 anyway.



git-svn-id: svn://svn.code.sf.net/p/octave/code/trunk/octave-forge@12544 fc592f45-0881-4044-91ff-c96495a54512